requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.22
paho-mqtt==2.1.0
jinja2==3.1.2
aiosqlite==0.19.0
python-multipart==0.0.6
